refactor: move prompt expansion
