+++ src/prompt_expand.h
@@ -1,0 +1,6 @@
+#ifndef PROMPT_EXPAND_H
+#define PROMPT_EXPAND_H
+
+char *expand_prompt(const char *prompt);
+
+#endif /* PROMPT_EXPAND_H */
