+++ src/repl.c
@@ -9,7 +9,7 @@
 #include "execute.h"
 #include "history.h"
 #include "history_expand.h"
-#include "var_expand.h"
+#include "prompt_expand.h"
 #include "lineedit.h"
 #include "jobs.h"
 #include "trap.h"
